modif

+++ ft_wstr.c
@@ -6,7 +6,7 @@
 /*   By: vbranco <marvin@le-101.fr>                 +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/01/24 18:44:46 by vbranco      #+#   ##    ##    #+#       */
-/*   Updated: 2018/01/24 20:24:26 by vbranco     ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/02/08 16:46:30 by vbranco     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -17,7 +17,7 @@ void		ft_wstr(wchar_t *wstr, char *str, t_form *form)
 {
 	size_t	i;
 	char	*tmp;
-	int	prec;
+	int		prec;
 
 	prec = form->prec;
 	i = 0;
@@ -25,7 +25,7 @@ void		ft_wstr(wchar_t *wstr, char *str, t_form *form)
 	{
 		tmp = ft_memalloc(65);
 		ft_wchar(wstr[i], tmp);
-		if (ft_strlen(tmp) > prec)
+		if ((int)ft_strlen(tmp) > prec)
 			ft_strcat(str, "\0");
 		else
 			ft_strcat(str, tmp);
